schools.sql
Completely drop your existing schools table then import the sql file into the learnandhelp database.

This file includes all of my dummy school data and the following constraints:

ALTER TABLE `schools`
  ADD CONSTRAINT CHK_type CHECK (type IN ('primary school', 'upper primary school', 'high school', 'other'));
ALTER TABLE `schools`
  ADD CONSTRAINT CHK_category CHECK (category IN ('private', 'public', 'other'));
ALTER TABLE `schools`
  ADD CONSTRAINT CHK_designation CHECK (contact_designation IN ('teacher', 'head master', 'volunteer', 'other'));
ALTER TABLE `schools`
  ADD CONSTRAINT CHK_status CHECK (status IN ('proposed', 'rejected', 'approved', 'completed'));
ALTER TABLE `schools`
  ADD CONSTRAINT CHK_start CHECK (grade_level_start > 0 AND grade_level_start < 11);
ALTER TABLE `schools`
  ADD CONSTRAINT CHK_end CHECK (grade_level_end > 0 AND grade_level_end < 11);

The constraints do not get exported when exporting the table data, they get stored in the information_schema --> CHECK_CONSTRAINTS table.  I verified them when adding the schools table in the assignments and they prevent adding a type/category/designation/status that is not in the listed values as well as preventing a starting or ending grade level that is not in the range of 1 to 10.