Update school_details.php
School details page was not showing the default school image.  Modified the code to correct that.  Per discussion with customer (the instructor) additional images will need to be tiled below the details.  Once that is  done the default image may be moved down below the details depending on how the page looks with the additional tiled images.  It may be preferable to just tile all of the images below the detail information.