Update admin_fill.php
Modified the code to use capitalized words in the dropdown lists instead of all lower case.  Will check for both capitalized first letters or all lower case (since some values in database are currently all lower case) to highlight which value is currently selected but will present everything with the first letter of each word capitalized and save it that way from now on.