Bug fixes, updates to adding a school
admin_delete_school.php - syntax errors in the code to delete media for the school, wasn't parsing properly

admin_edit_school.php - when adding a school the media tiling and upload button should not be visible

form_submit_school.php - after adding a school get the id of the record that was inserted then uses a hidden form to pass that to the admin_edit_school.php rather than returning to the list of schools.  User can then make any corrections needed and/or upload media for the new school.