Fixed password_change.php
made sure that the user inputting their email is actually part of the database. So if a user inputs a non-valid email an alert will show up telling them to try again. If a user inputs a valid email an alert will tell them an email has been sent and will be taken to a notification page.