Update Schools.php V2
Updated the code once more so now the image banner is fixed so if there is no images in the banners_images folder it will default use the image from the admin_icons folder as the banner image. Also implemented Sarah google pins code to categorize the google pins now.